pyproject.toml
[tool.poetry]
name = "zd"
version = "0.0.1"
description = "Find information for books, movies and music"
authors = ["Jianwei Han <hanjianwei@gmail.com>"]
repository = "https://github.com/hanjianwei/zd"
documentation = "https://hanjianwei.github.io/zd/"
readme = "README.md"
packages = [
  {include = "zd"}
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
click = "^8.1.3"
httpx = { extras = ["socks"], version = "^0.23.3" }
parsel = "^1.7.0"
pydantic = "^2.5.3"
loguru = "^0.7.2"
hishel = "^0.0.22"
python-frontmatter = "^1.1.0"
devtools = "^0.12.2"
pathvalidate = "^3.2.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
nox = "^2023.4.22"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
ruff = "^0.1.14"

[tool.poetry.scripts]
zd = "zd.cli:main"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "default"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["zd"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["zd"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
